Module 4 Completed

+++ Module 4 Tasks/3) fibonaci series.c
@@ -1,10 +1,10 @@
 #include<stdio.h>
 void main()
 {
-    long int a=1,b=1,c=2;
+    long long int a=1,b=1,c=2;
     for(int i=1;i<75;i++)
     {
-        printf("%lu\n",c);
+        printf("%lli\n",c);
         a=b;
         b=c;
         c=a+b;
